sept release (no spotify, &fixes)
[-] added dupli button in maindb
[x] fix import current month
[-] File Name should be standardised..
[x] add warning if lyrics added not to forget to select add new arangement
[-] add auto groups
[-] add maybe chery love to favs
[-] at import get youtube rksmith link (unsolved bug: async stays in waiting)
[-] at import get youtube song link (unsolved bug: async stays in waiting)
[-] compare/dupli manag 2-x selected songs
[-] add java 64bit if windows is for 64b https://www.java.com/en/download/manual.jsp
[-] 1st feeling inc. not coverting
[-] add better initial setup options & messages
[-] replaced way to call DDC to latest standard
[-] mamas and papas bonus (Added Part Field)
[-] mMa nd pPAS VOCALS (1st lyris ismissin)
[-] mMa nd pPAS VOCALS (1st lyric has the same start at second ()
[-] instead of old and new maybe add timestamp
[-] add song lenght in the duplicate management window
[-] tomahawk is not imported w the right version (same hash; any update/overrite of a song updates also original_filehash)
[-] search again should work if returning nada
[x] dlc name should only be unique after all dupli have been assesed
[x] cassius breaks (added stauff to lyrics from 0.o01 and dynamically size them)
[x] fix duplicate songs to incl new fields
[x] fix add vocals with new path
[x] if gearlist is NOT there disable use of db arg&tones values
[-] add dificulty field in arrangements
[-] check the in sort titles (re-enable it as lost in the single pack to multipack functions)
[-] clean lyrics doesnt clear some end empty lines
[-] renablin 1 and 2 option
[-] fascination stret bass is not properly removed (check if it really has DD as if the bonus/2nd track has the flag is set)
[-] love cats remove bass removal has issues (check if it really has DD as if the bonus/2nd track has the flag is set)
[x] change from pc to ps3 loses the ftp address (save only if PS3 is last format)
[x] fix for delete song incl also gearlist
[-] duplicate remov should clean title also for alt 1 alt 2
[-] added use internal logic flag per song
[x] fixe for packing with arang&tones details (amp and cab settings)
[x] fix for using internal logic not being selected

+++ RocksmithToolkitLib/DLCPackage/Packer.cs
@@ -76,6 +76,7 @@ namespace RocksmithToolkitLib.DLCPackage
         /// <param name="predefinedPlatform">Predefined source platform.</param>
         public static string Unpack(string sourceFileName, string savePath, bool decodeAudio = false, bool overwriteSongXml = false, Platform predefinedPlatform = null)
         {
+            var ps4 = false;
             Platform platform = sourceFileName.GetPlatform();
             if (predefinedPlatform != null && predefinedPlatform.platform != GamePlatform.None && predefinedPlatform.version != GameVersion.None)
                 platform = predefinedPlatform;
@@ -113,6 +114,7 @@ namespace RocksmithToolkitLib.DLCPackage
                     break;
                 case GamePlatform.PS3:
                     UnpackPS3Package(sourceFileName, savePath, platform);
+                    if (platform.version == GameVersion.None) { platform.version = GameVersion.RS2014; ps4 = true; } //bcapi
                     break;
                 case GamePlatform.None:
                     throw new InvalidOperationException("Platform not found :(");
@@ -124,8 +126,8 @@ namespace RocksmithToolkitLib.DLCPackage
             // DECODE AUDIO
             if (decodeAudio)
             {
-                GlobalExtension.ShowProgress("Decoding Audio ...", 50); //bcapi removing oggg as giving errors e.g. pretender official
-                var audioFiles = Directory.EnumerateFiles(unpackedDir, "*.*", SearchOption.AllDirectories).Where(s => s.EndsWith(".odgg") || s.EndsWith(".wem"));
+                GlobalExtension.ShowProgress("Decoding Audio ...", 50);
+                var audioFiles = Directory.EnumerateFiles(unpackedDir, "*.*", SearchOption.AllDirectories).Where(s => s.EndsWith(".ogg") || s.EndsWith(".wem"));
                 foreach (var file in audioFiles)
                 {
                     var outputAudioFileName = Path.Combine(Path.GetDirectoryName(file), String.Format("{0}_fixed{1}", Path.GetFileNameWithoutExtension(file), ".ogg"));
@@ -166,51 +168,54 @@ namespace RocksmithToolkitLib.DLCPackage
                             att = Manifest2014<Attributes2014>.LoadFromFile(jsonFiles).Entries.ToArray()[0].Value.ToArray()[0].Value;
                     }
 
-                    var sngContent = Sng2014File.LoadFromFile(sngFile, platform);
-                    using (var outputStream = new FileStream(xmlSngFile, FileMode.Create, FileAccess.ReadWrite))
+                    if (!ps4) //bcapi
                     {
-                        dynamic xmlContent = null;
-
-                        if (arrType == ArrangementType.Vocal)
-                            xmlContent = new Vocals(sngContent);
-                        else
-                            xmlContent = new Song2014(sngContent, att);
-
-                        xmlContent.Serialize(outputStream);
-                    }
-
-                    // correct old toolkit/EOF xml (tuning) issues ... sync with SNG data
-                    if (File.Exists(xmlEofFile) && !overwriteSongXml && arrType != ArrangementType.Vocal)
-                    {
-                        var eofSong = Song2014.LoadFromFile(xmlEofFile);
-                        var sngSong = Song2014.LoadFromFile(xmlSngFile);
-                        if (eofSong.Tuning != sngSong.Tuning)
+                        var sngContent = Sng2014File.LoadFromFile(sngFile, platform);
+                        using (var outputStream = new FileStream(xmlSngFile, FileMode.Create, FileAccess.ReadWrite))
                         {
-                            eofSong.Tuning = sngSong.Tuning;
-                            var xmlComments = Song2014.ReadXmlComments(xmlEofFile);
+                            dynamic xmlContent = null;
 
-                            using (var stream = File.Open(xmlEofFile, FileMode.Create))
-                                eofSong.Serialize(stream, true);
+                            if (arrType == ArrangementType.Vocal)
+                                xmlContent = new Vocals(sngContent);
+                            else
+                                xmlContent = new Song2014(sngContent, att);
 
-                            Song2014.WriteXmlComments(xmlEofFile, xmlComments, customComment: "Synced with SNG file");
-                            Console.WriteLine("Fixed Tuning Descrepancies: " + xmlEofFile);
-                            GlobalExtension.ShowProgress("Fixed tuning descepancies ...");
+                            xmlContent.Serialize(outputStream);
                         }
-                    }
-                    else if (File.Exists(xmlEofFile) && !overwriteSongXml && arrType == ArrangementType.Vocal)
-                    {
-                        // preserves vocal xml comments
-                    }
-                    else
-                    {
-                        if (!isODLC)
-                            Song2014.WriteXmlComments(xmlSngFile, customComment: "Generated from SNG file");
 
-                        File.Copy(xmlSngFile, xmlEofFile, true);
-                    }
+                        // correct old toolkit/EOF xml (tuning) issues ... sync with SNG data
+                        if (File.Exists(xmlEofFile) && !overwriteSongXml && arrType != ArrangementType.Vocal)
+                        {
+                            var eofSong = Song2014.LoadFromFile(xmlEofFile);
+                            var sngSong = Song2014.LoadFromFile(xmlSngFile);
+                            if (eofSong.Tuning != sngSong.Tuning)
+                            {
+                                eofSong.Tuning = sngSong.Tuning;
+                                var xmlComments = Song2014.ReadXmlComments(xmlEofFile);
 
-                    if (File.Exists(xmlSngFile))
-                        File.Delete(xmlSngFile);
+                                using (var stream = File.Open(xmlEofFile, FileMode.Create))
+                                    eofSong.Serialize(stream, true);
+
+                                Song2014.WriteXmlComments(xmlEofFile, xmlComments, customComment: "Synced with SNG file");
+                                Console.WriteLine("Fixed Tuning Descrepancies: " + xmlEofFile);
+                                GlobalExtension.ShowProgress("Fixed tuning descepancies ...");
+                            }
+                        }
+                        else if (File.Exists(xmlEofFile) && !overwriteSongXml && arrType == ArrangementType.Vocal)
+                        {
+                            // preserves vocal xml comments
+                        }
+                        else
+                        {
+                            if (!isODLC)
+                                Song2014.WriteXmlComments(xmlSngFile, customComment: "Generated from SNG file");
+
+                            File.Copy(xmlSngFile, xmlEofFile, true);
+                        }
+
+                        if (File.Exists(xmlSngFile))
+                            File.Delete(xmlSngFile);
+                    }
 
                     progress += step;
                     GlobalExtension.UpdateProgress.Value = (int)progress;
@@ -547,9 +552,10 @@ namespace RocksmithToolkitLib.DLCPackage
 
             var outputMessage = RijndaelEncryptor.DecryptPS3Edat();
 
-            if (File.Exists(outputFilename)) File.Delete(outputFilename);
+            if (File.Exists(outputFilename))
+                ;//File.Delete(outputFilename);
 
-            foreach (var fileName in Directory.EnumerateFiles(PS3_WORKDIR, "*.psarc.dat"))
+            foreach (var fileName in Directory.EnumerateFiles(PS3_WORKDIR, "*.psarc.edat"))
             {
                 using (var outputFileStream = File.OpenRead(fileName))
                 {
